Conflicts:
	.gradle/2.14.1/taskArtifacts/cache.properties.lock
	.gradle/2.14.1/taskArtifacts/fileHashes.bin
	.gradle/2.14.1/taskArtifacts/fileSnapshots.bin
	.gradle/2.14.1/taskArtifacts/fileSnapshotsToTreeSnapshotsIndex.bin
	.gradle/2.14.1/taskArtifacts/taskArtifacts.bin
	.gradle/2.14.1/tasks/_app_compileDebugJavaWithJavac/localClassSetAnalysis/localClassSetAnalysis.bin
	.gradle/2.14.1/tasks/_app_compileDebugJavaWithJavac/localClassSetAnalysis/localClassSetAnalysis.lock
	.gradle/2.14.1/tasks/_app_compileDebugJavaWithJavac/localJarClasspathSnapshot/localJarClasspathSnapshot.bin
	.gradle/2.14.1/tasks/_app_compileDebugJavaWithJavac/localJarClasspathSnapshot/localJarClasspathSnapshot.lock
	.idea/workspace.xml
	local.properties

+++ app/src/main/java/com/example/trickyalarm/AlarmReceiver.java
@@ -20,7 +20,7 @@ public class AlarmReceiver extends BroadcastReceiver {
     @Override
     public void onReceive(Context context, Intent intent) {
         //PowerManager powerManager = (PowerManager) context.getSystemService(context.POWER_SERVICE);
-        //PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "MyWakelockTag");
+        //PowerManager.WakeLock wakeLock = powerManager.newWakeLock(PowerManager.PARTIAL_WAKE_LOCK, "");
         //wakeLock.acquire();
         Intent intent1 = new Intent(context, SignalActivity.class);
         intent1.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
@@ -33,7 +33,7 @@ public class AlarmReceiver extends BroadcastReceiver {
      * @param context
      * @param time in which alarm should work.
      */
-    public void setAlarm(Context context, int time) {
+    public void setAlarm(Context context, long time) {
         AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
         Intent intent = new Intent(context, AlarmReceiver.class);
         PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, intent, 0);
